Migrate more player code to character (#50474)
