+++ src/player.h
@@ -128,43 +128,8 @@ class player : public Character
 
         // ranged.cpp
 
-        /**
-         * Check if a given body part is immune to a given damage type
-         *
-         * This function checks whether a given body part cannot be damaged by a given
-         * damage_unit.  Note that this refers only to reduction of hp on that part. It
-         * does not account for clothing damage, pain, status effects, etc.
-         *
-         * @param bp: Body part to perform the check on
-         * @param dam: Type of damage to check for
-         * @returns true if given damage can not reduce hp of given body part
-         */
-        bool immune_to( const bodypart_id &bp, damage_unit dam ) const;
-        /** Modifies a pain value by player traits before passing it to Creature::mod_pain() */
-        void mod_pain( int npain ) override;
-        /** Sets new intensity of pain an reacts to it */
-        void set_pain( int npain ) override;
-        /** Returns perceived pain (reduced with painkillers)*/
-        int get_perceived_pain() const override;
-
-        /** Knocks the player to a specified tile */
-        void knock_back_to( const tripoint &to ) override;
-
-        /** Returns multiplier on fall damage at low velocity (knockback/pit/1 z-level, not 5 z-levels) */
-        float fall_damage_mod() const override;
-        /** Deals falling/collision damage with terrain/creature at pos */
-        int impact( int force, const tripoint &pos ) override;
-
-        /** Returns overall % of HP remaining */
-        int hp_percentage() const override;
-
         /** Returns list of rc items in player inventory. **/
         std::list<item *> get_radio_items();
-
-        /** Siphons fuel (if available) from the specified vehicle into container or
-         * similar via @ref game::handle_liquid. May start a player activity.
-         */
-        void siphon( vehicle &veh, const itype_id &desired_liquid );
 
         /** Used for eating object at a location. Removes item if all of it was consumed.
         *   @returns trinary enum NONE, SOME or ALL amount consumed.
@@ -176,7 +141,6 @@ class player : public Character
          */
         trinary consume( item &target, bool force = false, bool refuel = false );
 
-        int get_lift_assist() const;
 
         bool list_ammo( const item &base, std::vector<item::reload_option> &ammo_list,
                         bool empty = true ) const;
